Methods - EnergyModelsHeat

Index

EnergyModelsHeat.dh_fraction

EnergyModelsHeat.dh_upgrade

EnergyModelsHeat.fraction_equal_mass

EnergyModelsHeat.is_heat

EnergyModelsHeat.t_return

EnergyModelsHeat.upgrade_equal_mass

EnergyModelsHeat.upgradeable_fraction

Extension methods

Constraint methods

Utility methods

EnergyModelsHeat.dh_fraction — Function

dh_fraction

Return fraction of surpus heat that can be used for district heating. Dispatch on HeatExchangerAssumptions when calculating d_fraction

EnergyModelsHeat.dh_upgrade — Function

dh_upgrade

Return needed power to upgrade to outflow of useable (for district heating) heat

EnergyModelsHeat.fraction_equal_mass — Function

fraction_equal_mass(pd::PinchData)

Calculate fraction of heat available for district heating at pinch point T_DH_cold

EnergyModelsHeat.is_heat — Function

is_heat(r)

Returns true for heat resources, false otherwise. Extend this by dispatching on the type for any alternative heat resource type implemented.

EnergyModelsHeat.t_return — Function

t_return(rh::ResourceHeat)
t_return(rh::ResourceHeat, t)

Return the return temperature defined for a ResourceHeat

EnergyModelsHeat.upgrade_equal_mass — Function

Assuming equal mass flows

EnergyModelsHeat.upgradeable_fraction — Function

upgradeable_fraction

Return fraction of surplus heat that can be upgraded