Methods - Accessing fields

Index

EnergyModelsHeat.cap_lower_bound

EnergyModelsHeat.driving_force_resource

EnergyModelsHeat.eff_carnot

EnergyModelsHeat.heat_in_resource

EnergyModelsHeat.heat_loss_factor

EnergyModelsHeat.pipe_length

EnergyModelsHeat.pipe_loss_factor

EnergyModelsHeat.resource_heat

EnergyModelsHeat.t_ground

EnergyModelsHeat.t_sink

EnergyModelsHeat.t_source

EnergyModelsHeat.t_supply

DHPipe

EnergyModelsHeat.pipe_length — Function

pipe_length(l::DHPipe)

Returns the length of the pipe l.

EnergyModelsHeat.pipe_loss_factor — Function

pipe_loss_factor(l::DHPipe)

Returns the heat loss factor λ for the pipe l.

EnergyModelsHeat.t_ground — Function

t_ground(l::DHPipe)

Returns the ground temperature for the pipe l.

EnergyModelsHeat.resource_heat — Function

resource_heat(l::DHPipe)

Returns the transported ResourceHeat for pipe l.

EnergyModelsHeat.t_supply — Function

t_supply(rh::ResourceHeat)
t_supply(rh::ResourceHeat, t)

Return the supply temperature defined for a ResourceHeat.

t_supply(l::DHPipe)

Returns the supply temperature of applied ResourceHeat.

HeatPump

EnergyModelsHeat.eff_carnot — Function

eff_carnot(n::HeatPump)
eff_carnot(n::HeatPump, t)

Returns the Carnot efficiency of heat pump n as TimeProfile or in operational period t.

EnergyModelsHeat.t_sink — Function

t_sink(n::HeatPump)
t_sink(n::HeatPump, t)

Returns the temperature of the heat sink for heat pump n as TimeProfile or in operational period t.

EnergyModelsHeat.t_source — Function

t_source(n::HeatPump)
t_source(n::HeatPump, t)

Returns the temperature of the heat source for heat pump n as TimeProfile or in operational period t.

EnergyModelsHeat.cap_lower_bound — Function

cap_lower_bound(n::HeatPump)

Returns the lower capacity bound for heat pump n.

EnergyModelsHeat.heat_in_resource — Function

heat_in_resource(n::HeatPump)

Returns the resource for heat input for heat pump n.

EnergyModelsHeat.driving_force_resource — Function

driving_force_resource(n::HeatPump)

Returns the resource for driving force, i.e., electricity, for heat pump n.

ThermalEnergyStorage

EnergyModelsHeat.heat_loss_factor — Function

heat_loss_factor(n::ThermalEnergyStorage)

Returns the heat loss factor for storage n.

HeatExchanger

PinchData